requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
pytz==2025.2
requests==2.32.5
RPi.GPIO==0.7.1
urllib3==2.5.0
